+++ src/components/myPage/CollectionItem.jsx
@@ -1,40 +1,23 @@
 import React, { useState } from 'react';
-import { Text, Accordion, Tooltip, Button, ThemeIcon, Flex, Badge } from '@mantine/core';
-import { DatePickerInput } from '@mantine/dates';
-import { IconLayersLinked, IconTrash, IconCalendar } from '@tabler/icons-react';
+import { Text, Accordion, Tooltip, Button, ThemeIcon, Flex } from '@mantine/core';
+import { IconLayersLinked, IconTrash } from '@tabler/icons-react';
 import { useRecoilValue } from 'recoil';
-import styled from '@emotion/styled';
-import { AccordionLabel } from './index';
+import { AccordionLabel, ModifiedDate } from './index';
 import { useDeleteUserContentMutation } from '../../hooks/mutations';
 import { userState, categoryState } from '../../recoil/atom';
 import { ActionIcons } from '../common';
-
-const EditButton = styled(Badge)`
-  cursor: pointer;
-`;
-
-const getAddedDate = modifiedAt => modifiedAt?.match(/^([a-zA-Z0-9_.+-]+)T/)[1].replace(/-/g, ' .');
 
 const CollectionItem = ({ item, setClicked, open }) => {
   const email = useRecoilValue(userState);
   const listName = useRecoilValue(categoryState);
 
   const [hovered, setHovered] = useState(false);
-  const [editMode, setEditMode] = useState(false);
 
   const { mutate: deleteUserContent } = useDeleteUserContentMutation();
 
   const handleClick = item => {
     setClicked(item);
     open();
-  };
-
-  const handleEditDate = () => {
-    setEditMode(!editMode);
-  };
-
-  const handleDatePicker = e => {
-    e.stopPropagation();
   };
 
   const handleMouseEnter = () => setHovered(true);
@@ -65,31 +48,7 @@ const CollectionItem = ({ item, setClicked, open }) => {
         </Accordion.Control>
         <Accordion.Panel w="90%" ml={55} mt={-15} mb={20}>
           <Flex direction="column" align="flex-start" gap={3}>
-            <Flex align="center" gap={10}>
-              {editMode ? (
-                <DatePickerInput
-                  onClick={handleDatePicker}
-                  placeholder="감상한 날을 선택해주세요"
-                  size="xs"
-                  w={200}
-                  maw={400}
-                  icon={<IconCalendar size="1.1rem" stroke={1.5} />}
-                />
-              ) : (
-                <Text size="sm">{getAddedDate(item?.modified_at)}에 추가함</Text>
-              )}
-
-              {listName === 'history' &&
-                (!editMode ? (
-                  <EditButton size="sm" variant="outline" onClick={handleEditDate}>
-                    날짜 수정
-                  </EditButton>
-                ) : (
-                  <EditButton size="sm" variant="filled" onClick={handleEditDate}>
-                    수정 완료
-                  </EditButton>
-                ))}
-            </Flex>
+            <ModifiedDate date={item?.modified_at} />
 
             <div>
               <Tooltip label="더보기" position="bottom-end" withArrow withinPortal>
