+++ src/components/myPage/ModifiedDate.jsx
@@ -1,0 +1,56 @@
+import { useState } from 'react';
+import { useRecoilValue } from 'recoil';
+import styled from '@emotion/styled';
+import { DatePickerInput } from '@mantine/dates';
+import { Text, Flex, Badge } from '@mantine/core';
+import { IconCalendar } from '@tabler/icons-react';
+import { categoryState } from '../../recoil/atom';
+
+const EditButton = styled(Badge)`
+  cursor: pointer;
+`;
+
+const getAddedDate = modifiedAt => modifiedAt?.match(/^([a-zA-Z0-9_.+-]+)T/)[1].replace(/-/g, ' .');
+
+const ModifiedDate = ({ date }) => {
+  const category = useRecoilValue(categoryState);
+  const [editMode, setEditMode] = useState(false);
+
+  const handleEditDate = () => {
+    setEditMode(!editMode);
+  };
+
+  const handleDatePicker = e => {
+    e.stopPropagation();
+  };
+
+  return (
+    <Flex align="center" gap={10}>
+      {editMode ? (
+        <DatePickerInput
+          onClick={handleDatePicker}
+          placeholder="감상한 날을 선택해주세요"
+          size="xs"
+          w={200}
+          maw={400}
+          icon={<IconCalendar size="1.1rem" stroke={1.5} />}
+        />
+      ) : (
+        <Text size="sm">{getAddedDate(date)}에 추가함</Text>
+      )}
+
+      {category === 'history' &&
+        (!editMode ? (
+          <EditButton size="sm" variant="outline" onClick={handleEditDate}>
+            날짜 수정
+          </EditButton>
+        ) : (
+          <EditButton size="sm" variant="filled" onClick={handleEditDate}>
+            수정 완료
+          </EditButton>
+        ))}
+    </Flex>
+  );
+};
+
+export default ModifiedDate;
